alx_travel_app/requirements.txt
Django
djangorestframework
django-cors-headers
drf-yasg
django-environ
mysqlclient
celery
rabbitmq